+++ Assets/Scripts/Structs/EnemyInfo.cs
@@ -3,13 +3,13 @@
     [System.Serializable]
     public struct EnemyInfo
     {
-        public sbyte hitPoints;
+        public int hitPoints;
         public float attackRange;
         public float chaseSpeed;
         public float sightDistance;
         public float speed;
 
-        public EnemyInfo(sbyte hp, float attackRange, float chaseSpeed, float sightDistance, float speed)
+        public EnemyInfo(int hp, float attackRange, float chaseSpeed, float sightDistance, float speed)
         {
             hitPoints = hp;
             this.attackRange = attackRange;
@@ -21,7 +21,7 @@
         public static EnemyInfo operator+(EnemyInfo e1, EnemyInfo e2)
         {
             EnemyInfo newEnemyInfo = new EnemyInfo();
-            newEnemyInfo.hitPoints = (sbyte)(e1.hitPoints + e2.hitPoints);
+            newEnemyInfo.hitPoints = e1.hitPoints + e2.hitPoints;
             newEnemyInfo.attackRange = e1.attackRange + e2.attackRange;
             newEnemyInfo.chaseSpeed = e1.chaseSpeed + e2.chaseSpeed;
             newEnemyInfo.sightDistance = e1.sightDistance + e2.sightDistance;
